requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
pillow
pyyaml
streamlit
